requirements.txt
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.13.2
typing_extensions==4.14.0
Werkzeug==3.1.3
python-dateutil
